pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nova-api"
version = "0.1.0"
description = "Nova Agent API - A powerful agent-based API service using Clean + Hexagonal Architecture"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core dependencies
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Database and storage
    "redis>=5.0.0",
    # LLM and AI dependencies
    "langchain>=0.1.0",
    "langgraph>=0.0.30",
    "langchain-core>=1.0.0",
    # Phase 1: LiteLLM for multi-provider support
    "litellm>=1.50.0",
    # Phase 2: Tools
    "json-repair>=0.30.0",
    "readability-lxml>=0.8.1",
    "beautifulsoup4>=4.12.0",
    # Phase 4: Skills
    "pyyaml>=6.0",
    "python-frontmatter>=1.0.0",
    # Phase 5: CLI
    "typer>=0.12.0",
    "rich>=13.0.0",
    # Phase 6: Telegram
    "python-telegram-bot>=21.0",
    # Web and API
    "sse-starlette>=1.0.0",
    "python-dotenv>=1.0.0",
    # Pydantic AI for LLM validation
    "pydantic-ai>=0.0.14",
    # Code quality
    "ruff>=0.14.2",
    "pytest-asyncio>=1.2.0",
    "pytest>=8.4.2",
    "langgraph-cli[inmem]>=0.4.4",
    "langmem>=0.0.30",
    "psycopg[binary]>=3.2.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]

[tool.ruff]
line-length = 88
select = ["E", "F"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = ["pytest>=8.4.2"]

[project.scripts]
nova = "app.interfaces.cli.app:app"
